Refactoring. Move functions in module 'filter'.
