+++ js/main.js
@@ -1,6 +1,7 @@
 import {getData} from './server.js';
 import {createMap} from './map.js';
-import {onResetEvent, onFilterChange} from './form.js';
+import {onResetEvent} from './form.js';
+import {onFilterChange} from './filter.js';
 import {startValidation} from './validation.js';
 
 getData((offers) => {
